Now with 100% more camera!
Implemented camera and time, changed movement and particles to factor in
time.

+++ Demo/code/Bullet.h
@@ -15,7 +15,7 @@ public:
 
 protected:
 	sf::Vector2f mVelocity;
-	int mCounter;
+	float		 mCounter;
 };
 
 #endif
